requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
itsdangerous==2.2.0
pillow==10.4.0
